+++ src/components/molecules/empty-state/index.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import Icon from '../../atoms/icon'
 import Button from '../../atoms/button'
 import Link from '../../atoms/link'
-import { Heading1, Paragraph } from '../../atoms/typography'
+import { Header, Paragraph } from '../../atoms/typography'
 import { colors, spacing } from '../../../tokens'
 
 const EmptyState = props => {
@@ -18,7 +18,7 @@ const EmptyState = props => {
   }
   return (
     <EmptyState.Wrapper>
-      <Heading1>{props.title}</Heading1>
+      <Header size={1}>{props.title}</Header>
       <Icon name={props.icon} size={100} color={colors.icon.light} />
       <Paragraph>{props.text}</Paragraph> {helpLink}
       <Button primary icon={props.action.icon} onClick={props.action.method}>
